fix: circleci cache configuration (#129)
The CI configuration used cache, which was relying on the repo SHA to key cache usage from. This is not what we want. We want the nightly job to build fresh images nightly as upstream bits may have changed (base images, cnb-develop, etc.). Using job-local storage instead of cache simplifies the job and will probably sped things up a bit as well.