Remove redundant JVM buildpack dependencies
These buildpacks are already transitive dependencies of `heroku/java`,
and are not specified separately in the detection order.

Removing them does not stop someone from using those buildpacks
individually (they will still be in the final builder image), but reduces the
number of image URLs that must be updated each time a release occurs.

This came up in:
https://github.com/heroku/pack-images/pull/187#pullrequestreview-717771917

Fixes GUS-W-9672375.