Use shimmed Ruby buildpack instead of the prototype hybrid CNB (#282)
The previously deployed Ruby buildpack is based on the "hybrid" approach https://github.com/heroku/heroku-buildpack-ruby. Essentially Terence added support for CNB into the existing buildpack. This works, but it's not taking advantage of most of the features in CNB. It's really not too different than using the "shim" buildpack.

(The *shim* is a wrapper around any legacy/v2 buildpack that allows it to interface with CNB)

The idea is that we'll use the shim buildpack for now and switch over once a native CNB implementation in Rust is ready.

GUS-W-11461073