Add experimental Python Functions CNB to `heroku/builder:22` (#304)
Adds experimental support for Python Salesforce Functions to
`heroku/builder:22`, to support the Python Functions alpha/beta release.

For now, this buildpack only passes detection for Python functions
(projects that contain Python related files, such as `requirements.txt`,
and also contain a `project.toml` with a `com.salesforce.type` of
`"function"`). The buildpack is also only uploaded to S3 for now - later
it will be published to Docker Hub and the upstream CNB registry.

For any other Python projects this buildpack will not pass detection,
so this builder image will act as it did before (i.e.: not yet support
them).

This PR adds the `sf generate function` template Python function under
the examples directory, which is then built and tested with a health
check request (like the other functions examples) in CI.

See also:
https://github.com/heroku/buildpacks-python/compare/functions-experimental
https://github.com/heroku/buildpacks-python/actions/runs/3661415824
https://github.com/heroku/sf-functions-python

GUS-W-12171308.