Update lifecycle from v0.15.3 to v0.16.0 (#318)
To pick up recent fixes/features, including an improvement to the build
log output in the case where all buildpacks fail to detect. With those
build log improvements the `--verbose` option to `pack build` is no
longer required in CI here, and has been removed to make the rest of the
logs less noisy.

This update also does bring with it a new deprecation warning for usages
of old buildpack and platform API versions. These are currently
triggered by buildpacks using the cnb-shim, however, they are just
warnings and should be resolved separately rather than here, to avoid
adding additional risk to this change.

See:
https://github.com/buildpacks/lifecycle/releases/tag/v0.16.0
https://github.com/buildpacks/lifecycle/compare/v0.15.3...v0.16.0

GUS-W-12538046.