Make Heroku-18 builds fail with a skippable EOL warning (#336)
The Heroku-18 stack reached end-of-life on April 30th 2023, and as
of May 1st 2023, builds are no longer possible on the (default) non-CNB
Heroku build system, and the underlying stack images will no longer
 receive security updates (beyond whatever updates Canonical choose to
release in May). 

However, anyone using the `heroku/buildpacks:18` CNB builder image
may not realise that it's now EOL and soon to be insecure. They may also
not realise that the builder image has moved from the `heroku/buildpacks`
Docker repo to `heroku/builder`.

In order to raise awareness of both of these, an EOL buildpack has been
added, which fails the build with an EOL warning, unless the env var
`ALLOW_INSECURE_HEROKU_18_BUILDER` is set to `1` during the build.

This should affect very few users, since Heroku is not yet using CNBs by default,
and Salesforce Functions use the `heroku/builder:22` builder image instead.

This buildpack was written in bash, since:
- the buildpack functionality we need is extremely simple (checking an env
  var, printing a message and exiting 1 or 0)
- the buildpack (and other Heroku-18 related files) are going to be deleted
  in a few weeks anyway, so its not worth spending time getting a Rust build
  workflow working with the existing complex GitHub Actions workflow here

GUS-W-13143190.