Update to Pack 0.29.0
Since it contains the fix for:
https://github.com/buildpacks/pack/issues/1453

...which will mean we no longer constantly publish new builder
images even if nothing has changed. (Since now the SHA256 will
stay the same each time, causing the docker push to be a no-op.)

Normally we'd wait for the default Pack version to be updated in the
`buildpacks/github-actions/setup-pack` Action, however, it hasn't been
updated yet:
https://github.com/buildpacks/github-actions/pull/196#issuecomment-1536796572

Once the default version is updated again, we can switch back to using
it (and having Dependabot keep the Action up to date), instead of having to
hardcode versions here.

GUS-W-13204219.