Improve the naming of the functions related CI jobs and test fixtures
* The CI job "test-examples" was already functions-specific, but is now
  renamed to "test-functions" to make that clearer.
* The functions test fixtures have been grouped together in a
  functions directory.

This change make it easier to understand the CI job matrix from the
abbreviated names in the summary (there should be less truncation
since the `language` matrix parameter has shorter values than the
previous `examples` parameter).

In addition, it will simplify the future changes if/when we decide to
move functions support to its own builder tag.