Set the `DYNO` env var when running integration tests
The `DYNO` env var is now set when running the built containers during
this repo's integration tests. This means they more accurately reflect the
Heroku environment, which is important since some of the getting started
guides (eg Python's) use the presence of the `DYNO` env var to determine
whether to enable production mode or not.

In the Python getting started guide, if DYNO is not set, then the guide runs
in debug mode, which means static assets can be dynamically fetched from
the source, and so bypass the normal production assets serving mechanism.
By setting `DYNO` the integration tests will now validate that the production
assets have been generated (via the `manage.py collectstatic` command that
is now automatically run by the Python CNB).

GUS-W-14144906.