Skip registry login if not publishing to that specific registry
When new builder images are published on main, they can be
published to two different registries:
1. The public Docker repo on Docker Hub
2. A private internal registry

Previously we'd always log into both registries, even if that particular
job was only publishing to one of the two registries.

Several of our images are already only published to one of the two
registries, and soon virtually all of them will be (as part of upcoming
PRs to add new image variants) - meaning it makes even more sense
to skip login if it's not needed.