Simplify the example functions fixtures
The example functions fixtures in this repo are only needed to provide
a quick smoke test of building a function using the functions builder
image (primarily to ensure that the builder order grouping is correct).
All other CNB functionality should (and is) tested in the CNB or function
invoker itself.

This means the testing/linting related files in the fixtures is unused and
can be removed,

Doing so will prevent an endless stream of Dependabot security alert PRs
for Node.js dev/test dependencies (eg: #419, #420, #422), now that we have
committed the NPM lockfiles (xref #418).