Stop publishing non-functions images to the internal registry
The `heroku/buildpacks:20` and `heroku/builder:22` builders are now
only published to Docker Hub, rather than also being published to the
internal registry (under the `heroku-20:builder` and `heroku-22:builder`
tags respectively).

Kodon has not defaulted to these non-functions images from the internal registry
for some time, and as of heroku/kodon#410 these images (and the internal
registry) are no longer on the builder allow list, so it's not possible to manually
select them using `project.toml`.

This change does not affect the functions-only image, which continues to be
published to (and consumed by Kodon from) the internal registry.