Add deprecation warnings to legacy builder images (#429)
Adds deprecation warnings to the `heroku/builder-classic:22` and
`heroku/buildpacks:20` CNB builder images, to raise awareness that they
are no longer recommended.

The warnings have been added using a warnings buildpack, similar to the
approach used for the Heroku-18 EOL in:
https://github.com/heroku/cnb-builder-images/pull/336

For now these are just warnings, however, in the future will be changed
to an error (skippable via env var, like in the PR linked above).

Since these are warnings, the buildpack has been added to the end of each
order group, such that the message is at the end of the overall build
log, and so hopefully more visible to end users. Once the warning is
turned into an error, the buildpack should be moved to the start of the
order groups, so the build fails early for improved UX.

A separate mostly-copy-pasted buildpack was used for each builder, since
there is no other easy way to customise the message shown for each (and I
wanted to include the name of the deprecated builder in the warning
message, so users know what image name to grep for).

These buildpacks were written in bash (rather than in Rust, using
`libcnb.rs`), since:
- the buildpack functionality we need is extremely simple
- the buildpack is temporary (it will be deleted when we stop updating
  these builders in the future), so we don't need/want long term
  published images on CNB registry or to have to set up a Rust
  compilation/packaging step in this repo.

The buildpacks give zero lint warnings when checked with shellcheck
locally.

GUS-W-14194729.
GUS-W-14194736.