Improve documentation and description metadata
The repo README has been overhauled, to:
- Give more context as to what a builder is, and link to the upstream docs.
- Emphasise that these images are experimental and not yet used for
  "normal" Heroku builds.
- Make it easier to understand the difference between the image variants,
  including attributes such as run image and lifecycle version.
- Make the usage instructions more accessible for beginners.
- Give users breadcrumbs to follow when trying to work out where
  to file issues.

The "update lifecycle" GitHub Actions workflow has been modified so
that it now also modifies the lifecycle version in the README in addition
to that in the builder manifests.

Lastly, the description metadata used for each builder (that is shown by
`pack builder inspect`) has also been updated to mention deprecations
and bring the descriptions inline with that used upstream:
https://github.com/buildpacks/pack/blob/9c8223caef09a8f73b9c145c0ccf03ab704aac53/internal/builder/suggested_builder.go#L16-L23

GUS-W-14194789.