Update EOL message shown for legacy builders (#513)
Updates the EOL message to reflect that we're now stopping
updating these images. I've also removed the explicit builder
image name example from the message, since it will get out
of date once future builders are released.

The messaging has to be updated in this separate PR first, so we
can do one more release with the new messaging before
removing the publishing jobs in the next PR.

Towards #512.