Update status of `heroku/builder:24`
Now that the PHP CNB supports Heroku-24 (incl ARM64) and has been
added to `heroku/builder:24`, it means the `heroku/builder:24` builder
is now complete (or at least as complete as our other builder images).

As such, we can:
- update the descriptions to mention it supports all languages
- mark `heroku/builder:24` as recommended and use it in the
  example usage instructions

I also spotted a few other small improvements that could be made
to the README.

GUS-W-14686693.
GUS-W-14688185.