Stop setting `DYNO` when running the test apps in CI
Since this env var was added in #385 primarily for the Python
Getting Started Guide, but that guide:
- No longer requires that the env var be set in order to run in
  production mode, as of:
  https://github.com/heroku/python-getting-started/pull/251
- Now redirects to HTTPS if `DYNO` is set (which is not what we
  want for these tests, since there is no TLS cert configured), as of:
  https://github.com/heroku/python-getting-started/pull/253

Also, we're about to add Direwolf tests for CNBs/Fir, which will
test the guides in a true Heroku-like setting, so it makes more
sense for these smoke tests to more accurately test the local
`pack build` workflow instead - rather than a hybrid of both.

Lastly, I've added `-L` to the curl usages (which makes it follow
redirects), so that any issues with redirects are caught in CI.
(Such as the Python Getting Started Guide HTTP 301 redirecting
to an HTTPS URL with no cert, when the `DYNO` env var was
set.)

GUS-W-18093965.